pyproject.toml
[project]
name = "pet-tasks"
version = "0.1.0"
description = "Various educational tasks on different topics"
readme = "README.md"
requires-python = ">=3.14"

[dependency-groups]
main-libs = [
    "aiohttp==3.13.1",
    "asyncpg==0.31.0",
    "click==8.3.0",
    "fastapi[standard]==0.121.3",
    "langgraph==1.0.4",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "uvicorn[standard]==0.38.0"
]

[project.optional-dependencies]
dev-libs = [
    "mypy==1.18.2",
    "pytest==8.4.2",
    "pytest-mock==3.15.1",
    "ruff==0.14.2",
]

[tool.uv]
default-groups = ["main-libs"]

[tool.ruff]
src = ["src"]
preview = true
unsafe-fixes = true
line-length = 120
target-version = "py314"

[tool.ruff.lint]
preview = true
select = [
    "A",    # flake8-builtins
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "C90",  # maccabe
    "COM",  # flake8-commas
    "D",    # pydocstyle
    "DTZ",  # flake8-datetimez
    "E",    # pycodestyle
    "ERA",  # flake8-eradicate
    "EXE",  # flake8-executable
    "F",    # pyflakes
    "FBT",  # flake8-boolean-trap
    "FLY",  # pyflint
    "FURB", # refurb
    "G",    # flake8-logging-format
    "I",    # isort
    "ICN",  # flake8-import-conventions
    "ISC",  # flake8-implicit-str-concat
    "LOG",  # flake8-logging
    "N",    # pep8-naming
    "PERF", # perflint
    "PIE",  # flake8-pie
    "PL",   # pylint
    "PT",   # flake8-pytest-style
    "PTH",  # flake8-use-pathlib
    "Q",    # flake8-quotes
    "RET",  # flake8-return
    "RSE",  # flake8-raise
    "RUF",  # ruff
    "S",    # flake8-bandit
    "SIM",  # flake8-simpify
    "SLF",  # flake8-self
    "SLOT", # flake8-slots
    "T100", # flake8-debugger
    "TRY",  # tryceratops
    "UP",   # pyupgrade
    "W",    # pycodestyle
    "YTT",  # flake8-2020
]
ignore = [
    "A005",    # Module `types` is shadowing a Python builtin module
    "C901",    # is too complex (14 > 10)
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D105",    # Missing docstring in magic method
    "D106",    # Missing docstring in public nested class
    "D107",    # Missing docstring in `__init__`
    "D203",    # 1 blank line required before class docstring
    "D213",    # Multi-line docstring summary should start at the second line
    "RUF002",  # Docstring ambiguous.
    "RUF029",  # Function is declared `async`, but doesn't `await` or use `async` features.
]
unfixable = [
    "B", # Avoid trying to fix flake8-bugbear violations.
    "ERA", # Avoid deleting commented-out code
]
mccabe = { max-complexity = 10 }
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
flake8-tidy-imports = { ban-relative-imports = "all" }

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.ruff.lint.per-file-ignores]
"{tests/*.py,*/test/*,test_*.py}" = [
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
    "ARG001",  # Unused function argument
    "ARG002",  # Unused method argument
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D107",    # Missing docstring in `__init__`
    "ERA001",  # Found commented-out code
    "FBT001",  # Boolean-typed positional argument in function definition
    "PLC0415", # `import` should be at the top-level of a file
    "PLR0913", # Too many arguments to function call
    "PLR0917", # Too many positional arguments
    "PLR2004", # Magic value used in comparison
    "PTH108",  # `os.unlink()` should be replaced by `Path.unlink()`
    "S101",    # allow `assert` for tests
    "S311",    # Standard pseudo-random generators are not suitable for cryptographic purposes
]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.mypy]
python_version = "3.14"
mypy_path = "./src/"
allow_redefinition = false
check_untyped_defs = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true
ignore_errors = false
ignore_missing_imports = false
implicit_reexport = false
local_partial_types = true
namespace_packages = true
no_implicit_optional = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

plugins = [
    "pydantic.mypy"
]

exclude = [
    "tests/",
]

[[tool.mypy.overrides]]
module = [
     # missing library stubs or py.typed marker
    "asyncpg.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_typed =true
init_forbid_extra = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
pet-tasks = "pet_tasks.cli:cli"